requirements.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
opencv-python
pillow
pytesseract
#